requirements.txt
Django==3.1
djangorestframework===3.11.1
psycopg2==2.8.5
numpy==1.19.1
Pillow==7.2.0
face-recognition==1.3.0
celery==4.4.7
redis==3.5.3
six==1.15.0
djangorestframework_simplejwt==4.4.0

flake8==3.8.3